requirements.txt
autopep8==1.7.1
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
tomli==2.0.1
